Cozy navigation, settings screen, updated design, some improvements
Added the ability to select a gender to the page for adding a new person. Added a tabbar, and with it the settings screen. In the settings, you can enable notifications and select the time to receive daily notifications, switch the color scheme of the application, and also return to the onboarding screen. Also made a lot of minor design changes and fixed a few bugs.